Move Biq sections into directory
